Move compare_logs outside of process_replay
