+++ selfdrive/debug/run_process_on_route.py
@@ -2,7 +2,7 @@
 
 import argparse
 
-from selfdrive.test.process_replay.compare_logs import save_log
+from selfdrive.test.compare_logs import save_log
 from selfdrive.test.process_replay.process_replay import CONFIGS, replay_process
 from tools.lib.logreader import MultiLogIterator
 from tools.lib.route import Route
