+++ selfdrive/test/process_replay/model_replay.py
@@ -14,7 +14,7 @@ from common.transformations.camera import tici_f_frame_size, tici_d_frame_size
 from system.hardware import PC
 from selfdrive.manager.process_config import managed_processes
 from selfdrive.test.openpilotci import BASE_URL, get_url
-from selfdrive.test.process_replay.compare_logs import compare_logs, save_log
+from selfdrive.test.compare_logs import compare_logs, save_log
 from selfdrive.test.process_replay.test_processes import format_diff
 from system.version import get_commit
 from tools.lib.framereader import FrameReader
