+++ selfdrive/test/process_replay/test_processes.py
@@ -9,7 +9,7 @@ from typing import Any, DefaultDict, Dict
 
 from selfdrive.car.car_helpers import interface_names
 from selfdrive.test.openpilotci import get_url, upload_file
-from selfdrive.test.process_replay.compare_logs import compare_logs, save_log
+from selfdrive.test.compare_logs import compare_logs, save_log
 from selfdrive.test.process_replay.process_replay import CONFIGS, PROC_REPLAY_DIR, FAKEDATA, check_openpilot_enabled, replay_process
 from system.version import get_commit
 from tools.lib.filereader import FileReader
